gRPC-DB-Python/Module-Users-DB-gRPC/requirements.txt
attrs==22.1.0
greenlet==1.1.3.post0
grpcio==1.49.1
grpcio-tools==1.49.1
jsonschema==4.16.0
passlib==1.7.4
protobuf==4.21.7
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dotenv==0.21.0
six==1.16.0
toml==0.10.2
cryptography==39.0.0
